mapProjectAssets.test: fix shared mocks and timing issues
After adding a LWC to the DOM, it takes some amount of time
for its initialization (including child to parent event communication)
to finish. Waiting for a turn on the JS event queue is not long enough,
so we now wait a short time, about 10 milliseconds,
before proceeding with test assertions.

Some tests were also getting incorrect results because of shared mocks,
so now these mocks are reconstructed in beforeEach.

Finally, we remove the 'loads without crashing' test which is redundant
with the others now.