validate: throw an error if extractor returns invalid type
Throw an error for Express if a developer made a mistake
by defining an extractor function that does not have the
correct return type of string or undefined.

This avoids the easy mistake of returning the whole object
you are extracting from instead of just the argument.

The error shows in server logs and responds with "Internal Server Error"
in PROD.