assetsController.find: filters are now optional and non-positional
The prior code with each positional filter param in the find
function created bugs elsewhere where it was assumed
the first argument was projectId. This bug caused our
endpoints to be called with NaN and get incorrect results.

Now, filters are specified in a map from filter parameter
to specified value, so filter parameters are no longer positional
arguments and now can be specified in any order.