Query assets by location (#65)
* Added envelope location search

The envelope search returns the assets that are located within a
rectangle created from the bounds minLat, minLon, maxLat, maxLon.

* client tests and fixing query string

* Lat is Y, Lon is X

* Adding radius search

Users can give a lon,lat point and a radius and query for assets within
the circle

* Added new polygon search

Tests are in progress
Had to break other tests because of lat/lon y/x mixup. Will fix soon.

Made parameters required for geometry search

* polygonFind uses LINESTRING

fixed an error due to the way prepared statements inserts parameters by
using ST_GeomFromText with a linestring instead of trying to insert SQL
routines. (its a good thing this didn't work because that would be a
source of sql injection)

Changed geometrySearch/polygon from a POST back to GET since the error
had to do with the prepared statement rather than sending a body in a
GET.

* Fixed, now lat=y lon=x

* fix boundingbox lat=y lon=x

* Newline

* Removed redundant where clause

* Fixed merge comments